requirements.txt
nltk==3.5
numpy==1.19.2
tqdm==4.51.0
six==1.15.0
PyYAML==5.3.1
tensorflow~=2.2.0
transformers==3.4.0
requests~=2.24.0
scikit-learn~=0.23.2
torch~=1.6.0
torchtext~=0.6.0
pytorch-nlp
